+++ dune/istl/test/spqrtest.cc
@@ -48,7 +48,7 @@ void run(std::size_t N)
     mrs.insert(s);
 
   solver1.setSubMatrix(mat,mrs);
-  solver1.setVerbosity(true);
+  solver1.setVerbosity(1);
 
   solver.apply(x,b,res);
   solver.free();