+++ dune/istl/umfpack.hh
@@ -273,15 +273,15 @@ namespace Dune {
      * (and use a lot of memory).
      *
      *  @param matrix the matrix to solve for
-     *  @param verbose [0..2] set the verbosity level, defaults to 0
+     *  @param verbosity [0..2] set the verbosity level, defaults to 0
      */
-    UMFPack(const Matrix& matrix, int verbose=0) : matrixIsLoaded_(false)
+    UMFPack(const Matrix& matrix, int verbosity=0) : matrixIsLoaded_(false)
     {
       //check whether T is a supported type
       static_assert((std::is_same<T,double>::value) || (std::is_same<T,std::complex<double> >::value),
                     "Unsupported Type in UMFPack (only double and std::complex<double> supported)");
       Caller::defaults(UMF_Control);
-      setVerbosity(verbose);
+      setVerbosity(verbosity);
       setMatrix(matrix);
     }
 
@@ -291,15 +291,15 @@ namespace Dune {
      * (and use a lot of memory).
      *
      * @param matrix the matrix to solve for
-     * @param verbose [0..2] set the verbosity level, defaults to 0
+     * @param verbosity [0..2] set the verbosity level, defaults to 0
      */
-    UMFPack(const Matrix& matrix, int verbose, bool) : matrixIsLoaded_(false)
+    UMFPack(const Matrix& matrix, int verbosity, bool) : matrixIsLoaded_(false)
     {
       //check whether T is a supported type
       static_assert((std::is_same<T,double>::value) || (std::is_same<T,std::complex<double> >::value),
                     "Unsupported Type in UMFPack (only double and std::complex<double> supported)");
       Caller::defaults(UMF_Control);
-      setVerbosity(verbose);
+      setVerbosity(verbosity);
       setMatrix(matrix);
     }
 
@@ -329,20 +329,20 @@ namespace Dune {
     /** @brief Try loading a decomposition from file and do a decomposition if unsuccessful
      * @param mat_ the matrix to decompose when no decoposition file found
      * @param file the decomposition file
-     * @param verbose the verbosity level
+     * @param verbosity the verbosity level
      *
      * Use saveDecomposition(char* file) for manually storing a decomposition. This constructor
      * will decompose mat_ and store the result to file if no file wasn't found in the first place.
      * Thus, if you always use this you will only compute the decomposition once (and when you manually
      * deleted the decomposition file).
      */
-    UMFPack(const Matrix& mat_, const char* file, int verbose=0)
+    UMFPack(const Matrix& mat_, const char* file, int verbosity=0)
     {
       //check whether T is a supported type
       static_assert((std::is_same<T,double>::value) || (std::is_same<T,std::complex<double> >::value),
                     "Unsupported Type in UMFPack (only double and std::complex<double> supported)");
       Caller::defaults(UMF_Control);
-      setVerbosity(verbose);
+      setVerbosity(verbosity);
       int errcode = Caller::load_numeric(&UMF_Numeric, const_cast<char*>(file));
       if ((errcode == UMFPACK_ERROR_out_of_memory) || (errcode == UMFPACK_ERROR_file_IO))
       {
@@ -359,11 +359,11 @@ namespace Dune {
 
     /** @brief try loading a decomposition from file
      * @param file the decomposition file
-     * @param verbose the verbosity level
+     * @param verbosity the verbosity level
      * @throws Dune::Exception When not being able to load the file. Does not need knowledge of the
      * actual matrix!
      */
-    UMFPack(const char* file, int verbose=0)
+    UMFPack(const char* file, int verbosity=0)
     {
       //check whether T is a supported type
       static_assert((std::is_same<T,double>::value) || (std::is_same<T,std::complex<double> >::value),
@@ -376,7 +376,7 @@ namespace Dune {
         DUNE_THROW(Dune::Exception, "IO error while loading UMFPack decomposition");
       matrixIsLoaded_ = true;
       std::cout << "UMFPack decomposition successfully loaded from " << file << std::endl;
-      setVerbosity(verbose);
+      setVerbosity(verbosity);
     }
 
     virtual ~UMFPack()
@@ -793,8 +793,8 @@ namespace Dune {
       std::enable_if_t<
                 isValidBlock<typename Dune::TypeListElement<1, TL>::type::block_type>::value,int> = 0) const
     {
-      int verbose = config.get("verbose", 0);
-      return std::make_shared<Dune::UMFPack<M>>(mat,verbose);
+      int verbosity = config.get("verbosity", 0);
+      return std::make_shared<Dune::UMFPack<M>>(mat,verbosity);
     }
 
     // second version with SFINAE to validate the template parameters of UMFPack