+++ CHANGELOG.md
@@ -36,6 +36,8 @@ SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
 - Define `field_type` and `real_type` in `MultiTypeBlock[Matrix|Vector]` only if a common type of these
   types exist over all blocks in the container. Otherwise it is defined as `Std::nonesuch`.
 
+- Rename `[bool] verbose` parameter to `[int] verbosity`.
+
 ## Deprecations and removals
 
 - The deprecated CMake variables `SUPERLU_INCLUDE_DIRS`, `SUPERLU_LIBRARIES`,